20260520: keep filtering fix under challenge limit
