+++ tests/test_pbs_sample_regressions.py
@@ -54,6 +54,33 @@ def test_decide_remapping_handles_mixed_numeric_and_named_nodes():
     assert cluster.decide_remapping(["1", ""]) is True
 
 
+def test_filtering_name_patterns_removes_nodes_without_user_remap(monkeypatch):
+    args = SimpleNamespace(ANONYMIZE=False, BLINDREMAP=False, COLOR="OFF", REMAP=False)
+    monkeypatch.setattr(qtop, "args", args, raising=False)
+    monkeypatch.setattr(qtop, "dynamic_config", {}, raising=False)
+    monkeypatch.setattr(qtop, "user_to_color", {}, raising=False)
+    config = {"exotic_starting_wn_nr": 1000, "filtering": [{"exclude_name_patterns": ["wn002"]}], "percentage": 0.8, "remapping": [], "sorting": {}, "workernodes_matrix": [{"wn id lines": {"max_len": 0}}]}
+    nodes = [{"domainname": name, "state": [qtop.utils.ColorStr("-")], "np": 1, "core_job_map": {}} for name in ["wn001.example", "wn002.example", "wn003.example"]]
+    document = SimpleNamespace(jobs_dict={}, queues_dict={}, total_running_jobs=0, total_queued_jobs=0)
+
+    cluster = qtop.Cluster(document, nodes, qtop.WNFilter, config, args)
+
+    assert [node["domainname"] for node in cluster.workernode_dict.values()] == ["wn001.example", "wn003.example"]
+
+
+def test_dynamic_empty_filtering_does_not_force_remapping(monkeypatch):
+    monkeypatch.setattr(qtop, "dynamic_config", {"filtering": []}, raising=False)
+    cluster = qtop.Cluster.__new__(qtop.Cluster)
+    cluster.total_wn = 3
+    cluster.args = SimpleNamespace(BLINDREMAP=False)
+    cluster.node_subclusters = {"wn"}
+    cluster.workernode_list = [1, 2, 3]
+    cluster.offdown_nodes = 0
+    cluster.config = {"exotic_starting_wn_nr": 1000, "filtering": [{"exclude_name_patterns": ["wn002"]}], "percentage": 0.8}
+
+    assert cluster.decide_remapping(["1", "2", "3"]) is False
+
+
 def test_valid_corejobs_skips_jobs_missing_from_qstat():
     occupancy = qtop.WNOccupancy.__new__(qtop.WNOccupancy)
